requirement.txt
abnf==1.1.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
certifi==2023.5.7
charset-normalizer==3.2.0
cytoolz==0.12.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-abi==4.1.0
eth-account==0.8.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
parsimonious==0.9.0
Pillow==10.0.0
protobuf==4.23.4
pycryptodome==3.18.0
pydantic==1.10.11
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunormalize==15.0.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rfc3987==1.3.8
rlp==3.0.0
rpds-py==0.9.2
siwe==2.2.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.4
web3==6.6.1
websockets==11.0.3
yarl==1.9.2
